--- keyboards/rgbkb/pan/rev1/rev1.c
@@ -1,42 +1,0 @@
-#include "quantum.h"
-#include "rgb_matrix_types.h"
-
-#if defined(RGB_ENCODER_LEFT)
-#endif
-#if defined(RGB_ENCODER_RIGHT)
-#endif
-
-#if defined(RGB_MATRIX_ENABLE)
-    led_config_t g_led_config = { {
-       { 0,  1,  2,  3,  4,  5,  6,  7,  8,  9,  10, 11, 12 },
-       { NO_LED, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14, 13 },
-       { 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, NO_LED  },
-       { 50, 49, 48, 47, 46, 45, 44, 43, 42, 41, 40, 39, 38 },
-       { 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, NO_LED  }
-    }, {
-#if defined(RGB_ENCODER_LEFT)
-    {0, 0},
-#endif
-#if defined(RGB_ENCODER_RIGHT)
-    {228, 0},
-#endif
-        {0, 0}, {19, 0}, {38, 0}, {57, 0}, {76, 0}, {95, 0}, {114, 0}, {133, 0}, {152, 0}, {171, 0}, {190, 0}, {209, 0}, {228, 0} ,
-        {238, 13}, {219, 13}, {200, 13}, {181, 13}, {162, 13}, {143, 13}, {124, 13}, {105, 13}, {86, 13}, {67, 13}, {48, 13}, {29, 13}, 
-        {0, 26}, {19, 26}, {38, 26}, {57, 26}, {76, 26}, {95, 26}, {114, 26}, {133, 26}, {152, 26}, {171, 26}, {190, 26}, {209, 26},
-        {247, 38}, {228, 38}, {209, 38}, {190, 38}, {171, 38}, {152, 38}, {133, 38}, {114, 38}, {95, 38}, {76, 38}, {57, 38}, {38, 38}, {19, 38},
-        {0, 51}, {19, 51}, {38, 51}, {57, 51}, {76, 51}, {95, 51}, {114, 51}, {133, 51}, {152, 51}, {171, 51}, {190, 51}, {209, 51},
-    }, {
-#if defined(RGB_ENCODER_LEFT)
-    0,
-#endif
-#if defined(RGB_ENCODER_RIGHT)
-    0,
-#endif
-        1, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 1,
-        1, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 1,
-        1, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 1,
-        1, 4, 4, 4, 4, 4, 4, 4, 4, 4, 4, 1, 1,
-        1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1
-    } };
-#endif
-
