--- keyboards/rgbkb/pan/rev1/rev1.h
@@ -1,1 +1,0 @@
-#pragma once
